updated test classes names
