+++ src/test/java/sfa/SFAWordsVariableLengthTest.java
@@ -4,6 +4,7 @@ package sfa;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
 
 import org.junit.Test;
 import org.junit.runners.JUnit4;
@@ -19,10 +20,11 @@ import sfa.transformation.SFA.HistogramType;
  *
  */
 @RunWith(JUnit4.class)
-public class SFAWords {
+public class SFAWordsVariableLengthTest {
 
     @Test
-    public void testSFAWords() throws IOException {
+    public void testSFAWordsVarLength() throws IOException {
+
         int symbols = 8;
         int wordLength = 16;
         boolean normMean = true;
@@ -30,7 +32,7 @@ public class SFAWords {
         SFA sfa = new SFA(HistogramType.EQUI_DEPTH);
 
         // Load the train/test splits
-        ClassLoader classLoader = SFAWords.class.getClassLoader();
+        ClassLoader classLoader = SFAWordsTest.class.getClassLoader();
         TimeSeries[] train = TimeSeriesLoader
                              .loadDatset(new File(classLoader.getResource("datasets/CBF/CBF_TRAIN")
                                          .getFile()));
@@ -38,7 +40,7 @@ public class SFAWords {
                             .loadDatset(new File(classLoader.getResource("datasets/CBF/CBF_TEST")
                                         .getFile()));
 
-        // train SFA representation
+        // train SFA representation using wordLength
         sfa.fitTransform(train, wordLength, symbols, normMean);
 
         // bins
@@ -47,17 +49,19 @@ public class SFAWords {
         // transform
         for (int q = 0; q < test.length; q++) {
             short[] wordQuery = sfa.transform(test[q]);
-            System.out.println("Time Series " + q + "\t" + toSfaWord(wordQuery));
+
+            // iterate variable lengths
+            for (int length = 4; length <= wordLength; length*=2) {
+                System.out.println("Time Series " + q + "\t" + length + "\t" + toSfaWord(Arrays.copyOf(wordQuery, length)));
+            }
         }
     }
 
     public static String toSfaWord(short[] word) {
         StringBuffer sfaWord = new StringBuffer();
-
         for (short c : word) {
             sfaWord.append((char)(Character.valueOf('a') + c));
         }
-
         return sfaWord.toString();
     }
 }
